--- src/main/java/org/example/oddventure/admin/controller/AdminController.java
@@ -1,11 +1,0 @@
-package org.example.oddventure.admin.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/admin")
-public class AdminController {
-}
