--- src/main/java/org/example/oddventure/admin/service/AdminService.java
@@ -1,11 +1,0 @@
-package org.example.oddventure.admin.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-@RequiredArgsConstructor
-@Transactional
-public class AdminService {
-}
