--- src/main/java/org/example/oddventure/bet/controller/BetController.java
@@ -1,9 +1,0 @@
-package org.example.oddventure.bet.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class BetController {
-}
