--- src/main/java/org/example/oddventure/bet/dto/BetRequest.java
@@ -1,4 +1,0 @@
-package org.example.oddventure.bet.dto;
-
-public record BetRequest() {
-}
