--- src/main/java/org/example/oddventure/bet/dto/BetResponse.java
@@ -1,4 +1,0 @@
-package org.example.oddventure.bet.dto;
-
-public record BetResponse() {
-}
