--- src/main/java/org/example/oddventure/bet/enums/SelectedTeam.java
@@ -1,6 +1,0 @@
-package org.example.oddventure.bet.enums;
-
-public enum SelectedTeam {
-    Team_A,
-    Team_B
-}
