--- src/main/java/org/example/oddventure/bet/service/BetService.java
@@ -1,9 +1,0 @@
-package org.example.oddventure.bet.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class BetService {
-}
