+++ src/main/java/org/example/oddventure/domain/bet/controller/BetController.java
@@ -1,0 +1,9 @@
+package org.example.oddventure.domain.bet.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class BetController {
+}
