+++ src/main/java/org/example/oddventure/domain/bet/dto/BetRequest.java
@@ -1,0 +1,4 @@
+package org.example.oddventure.domain.bet.dto;
+
+public record BetRequest() {
+}
