+++ src/main/java/org/example/oddventure/domain/bet/enums/SelectedTeam.java
@@ -1,0 +1,6 @@
+package org.example.oddventure.domain.bet.enums;
+
+public enum SelectedTeam {
+    Team_A,
+    Team_B
+}
