+++ src/main/java/org/example/oddventure/domain/bet/repository/BetRepository.java
@@ -1,0 +1,7 @@
+package org.example.oddventure.domain.bet.repository;
+
+import org.example.oddventure.domain.bet.entity.Bet;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface BetRepository extends JpaRepository<Bet, Long> {
+}
