+++ src/main/java/org/example/oddventure/domain/bet/service/BetService.java
@@ -1,0 +1,9 @@
+package org.example.oddventure.domain.bet.service;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class BetService {
+}
