+++ src/main/java/org/example/oddventure/domain/match/controller/MatchController.java
@@ -1,0 +1,52 @@
+package org.example.oddventure.domain.match.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.example.oddventure.common.dto.reponse.ApiPageResponse;
+import org.example.oddventure.common.dto.reponse.ApiResponse;
+import org.example.oddventure.domain.match.dto.response.MatchResponse;
+import org.example.oddventure.domain.match.entity.Match;
+import org.example.oddventure.domain.match.enums.MatchStatus;
+import org.example.oddventure.domain.match.service.MatchService;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/v1/matches")
+public class MatchController {
+
+    private final MatchService matchService;
+
+    @GetMapping
+    public ResponseEntity<ApiPageResponse<MatchResponse>> getMatches(
+            @RequestParam(defaultValue = "0") int page,
+            @RequestParam(defaultValue = "10") int size
+    ) {
+        Pageable pageable = PageRequest.of(page, size);
+
+        Page<MatchResponse> matches = matchService.getMatches(pageable);
+
+        return ApiPageResponse.success(matches);
+    }
+
+    @GetMapping("/{matchId}")
+    public ResponseEntity<ApiResponse<MatchResponse>> getMatch(
+            @PathVariable Long matchId
+    ) {
+
+        return ApiResponse.success(matchService.getMatch(matchId));
+    }
+
+    @GetMapping("/search")
+    public List<Match> searchMatches(
+            @RequestParam(required = false) String teamName,
+            @RequestParam(required = false) MatchStatus status
+    ) {
+        return matchService.searchMatches(teamName, status);
+    }
+}
