+++ src/main/java/org/example/oddventure/domain/match/enums/MatchStatus.java
@@ -1,0 +1,7 @@
+package org.example.oddventure.domain.match.enums;
+
+public enum MatchStatus {
+    SCHEDULED,
+    ONGOING,
+    FINISHED
+}
