+++ src/main/java/org/example/oddventure/domain/match/enums/MatchWinner.java
@@ -1,0 +1,7 @@
+package org.example.oddventure.domain.match.enums;
+
+public enum MatchWinner {
+    TEAM_A,
+    TEAM_B,
+    NO_MATCH
+}
