--- src/main/java/org/example/oddventure/match/enums/MatchStatus.java
@@ -1,7 +1,0 @@
-package org.example.oddventure.match.enums;
-
-public enum MatchStatus {
-    SCHEDULED,
-    ONGOING,
-    FINISHED
-}
