--- src/main/java/org/example/oddventure/match/enums/MatchWinner.java
@@ -1,7 +1,0 @@
-package org.example.oddventure.match.enums;
-
-public enum MatchWinner {
-    TEAM_A,
-    TEAM_B,
-    NO_MATCH
-}
